:recycle: separate message store #91
