+++ src/entities/message/index.ts
@@ -1,3 +1,5 @@
 export { MessageNotice } from './ui/message-notice';
 export { MessageReceived } from './ui/message-received';
 export { MessageSent } from './ui/message-sent';
+
+export { useMessageStore } from './store';
