+++ src/features/websocket/provider.tsx
@@ -2,6 +2,7 @@
 
 import { useEffect } from 'react';
 
+import { useMessageStore } from '@/entities/message';
 import { getToken } from '@/shared/api';
 
 import { useWebsocketStore } from './store';
@@ -11,8 +12,8 @@ export const WebsocketProvider: FC<{
   roomIds: string[];
 }> = ({ children, roomIds }) => {
   useEffect(() => {
-    const { client, setIsConnected, addNewMessage } =
-      useWebsocketStore.getState();
+    const { client, setIsConnected } = useWebsocketStore.getState();
+    const { addNewMessage } = useMessageStore.getState();
 
     getToken().then((token) => {
       if (token) {
